C runtime implementation for implSquareToLen([II[II)[I
This commit adds a C runtime function for the Java
intrinsic candidate implSquareToLen to improve throughput.
implSquareToLen is the underlying implementation of the squaring function
used by java.math.BigInteger multiplication when both
operands are equal.

Additionally, the parent caller methods implSquare and sibling callee
implSquareToLenChecks are made always worth inling because:

implSquare consists of two calls implSquareToLenChecks, implSquareToLen,
the latter of which is transformed to a helper call.

implSquareToLenChecks consists of 4 null and bound tests which should
always be inlined to the parent caller.

Signed-off-by: James You <james.you@protonmail.com>