Store J9VMThread->scopedValueCache in the Continuation object
When a virtual thread is mounted, the corresponding carrier thread
unmounts and no longer runs. Similarly, when a virtual thread is
unmounted, the corresponding carrier thread mounts and resumes its
operations.

When a virtual thread is mounted, the scopedValueCache for the carrier
thread is stored in the Continuation object's scopedValueCache field.

When a virtual thread is unmounted, the carrier thread's
scopedValueCache, which is stored in the Continuation object's
scopedValueCache field, is swapped with J9VMThread->scopedValueCache.

This assures a one to one mapping between a java.lang.Thread and
scopedValueCache. The live thread's scopedValueCache is always kept in
J9VMThread->scopedValueCache whereas the unmounted thread's
scopedValueCache is stored in the corresponding Continuation object's
scopedValueCache field.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>