Do not create class unload PIC site assumption if not required
The issue is `TR_RelocationRecordSymbolFromManager::activatePointer`
creates `TR_UnloadedClassPicSite` for a J9Method in the pre-prologue
of the current method body. The class of the J9Method is the same as
the class of the current method body. In this case, the class unload
assumption is not needed because once the class is unloaded, no code
will ever reach the J9Method in the pre-prologue.

Fixes: #17734

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>