Set isJitInduceOSRCall in x86 call snippet on 32-bit
Previously it would be set only on 64-bit. The only effect of leaving it
unset on 32-bit was to generate an unnecessary move instruction:

    mov edi, <helper-address>

immediately before the jump to the OSR induction helper, which does not
need to be passed its own address.

This change eliminates the unnecessary instruction on 32-bit and makes
the code more consistent.