Prevent prepareToFixMemberNames() from encountering dark matter
Previously it was possible for prepareToFixMemberNames() to come across
an ostensible MemberName object that was actually dead, and whose fields
could therefore be dangling pointers. Because prepareToFixMemberNames()
needs to dereference some of the fields, this could lead to a crash.

The crash that has been observed involved a dead MemberName that used to
represent a method of a class that had since been unloaded. Calling
getJNIMethodID() on the stale J9Method found what would have previously
been the correct J9Class, and then crashed for lack of a class loader.

Other similar crashes would have been possible as well. For example, if
the memory of the J9Method had been reused for something else, we could
have crashed in J9_CLASS_FROM_METHOD.

With this commit, class redefinition will run GC to eliminate any such
dead objects ("dark matter"), ensuring that prepareToFixMemberNames()
won't see them.