Fix a missing allocationFence in process_java_lang_StringUTF16_toBytes()
When we transform a StringUTF16.toBytes() into a newarray and a call to
String.decompressedArrayCopy(), we neglected to add an allocationFence.
This change will introduce an allocationFence after the
decompressedArrayCopy() call to ensure that all threads see the new
contents of memory on weak memory coherency machines like POWER and
AArch64.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>