Verify NullRestricted field classes meet requirements
NullRestricted fields must be a value type with an ImplicitCreation attribute that has the ACC_DEFAULT flag set. For non-static fields this should be checked during class creation and for static fields during the preparation stage of linking.

This pr also adds static NullRestricted fields to the flattenedClassCache to make the described failure possible. Changes to allow non-static NullRestricted fields to be flattened will come in a later pull request.

Signed-off-by: Theresa Mammarella <Theresa.T.Mammarella@ibm.com>