Fix reflect ConstantPool bootstrapping issues
We currently perform explicit native registration for
jdk.internal.reflect.ConstantPool. This occurs after JCL init in VM
startup. It is possible that ConstantPool will be used in JCL init before
the natives are registered, this will result in signal 218 since we have
public JNI stubs as a catch all for these capabilities.

I propose the following:

Step 1) (this PR)
- introduce a registerNative call in ConstantPool.

Step 2)
- update the JCL code to use the registerNative() in the static initializer
of the class, thereby ensuring that the natives are registered before the
class is used.

Step 3)
- remove the explicit register native call

With these steps we dont need to do any build coordination between openj9
and the extension repos.

Signed-off-by: tajila <atobia@ca.ibm.com>