Prevent deadlock due to unloading during class redefinition
When GC runs as part of class redefinition, it could unload classes and
therefore try to acquire the class unload mutex, which is already held
to prevent JIT compilation from running concurrently with redefinition.
Such an acquisition attempt could deadlock if it occurs on a different
thread, even though the acquisition would be morally recursive.

Class redefinition now sets a flag in the J9JavaVM to allow GC to detect
this situation and skip acquiring the mutex.