Adjust heap size for CRIU restore
When we snapshot on a machine with higher physical memory and restore on one
with lower physical memory, the InstantOn JVM heap size will be larger than
a usual JVM run on the restore machine. And the momory usage will be abnormally
higher.

Update the extension with the correct physical memory limit and utilize the
existing softmax mechanism to solve the problem. Note we only update the softmax
if it is larger than the minHeap size, smaller than the maxHeap size and
previously set softmax size (if any).

Related: #17596

Signed-off-by: Frank Kang <frank.kang@ibm.com>