Fix code cache segment race condition
When a new code cache needs to be allocated, first the VM allocates
a memory segment and then the JIT allocates a TR::CodeCache structure.
Finally, the JIT writes a pointer to the TR::CodeCache structure
at the beginning of the memory segment.

After the code cache segment has been added to the corresponding
VM segment list, but before the JIT had the chance to write the
TR::CodeCache pointer, a Java application thread could inquire
about memory consumption stats (with MXBeans) which could lead to the
VM reading a random value for the TR::CodeCache pointer and crash
 when dereferencing the bad pointer.

The solution is to write a NULL pointer at the begining of the code
cache segment as soon as that segment is allocated. Thus, a third
party could see either NULL (which is checked against by the reader)
or a valid TR::CodeCache pointer.

The solution assumes that reads/writes of pointers are atomic on
all supported platforms (i.e. 'tearing' is not possible when
the values to be written/read are naturaly aligned).

Signed-off-by: Marius <mpirvu@ca.ibm.com>