Guard the RecreateClassFileOnload option on the patchMap being NULL
This patch eliminates the assertion failures reported in issue
eclipse-openj9#14546. The reported failures only occur if (1) the build
uses JDK11 with OpenJDK methodhandles enabled; and (2) the class file
to be recreated has a non-null patch map. This patch has performance
implications in that the RecreateClassFileOnload optimization will no
longer apply to classes that have a non-null J9ClassPatchMap.

Closes: eclipse-openj9#14546
Signed-off-by: Nathan Henderson <nathan.henderson@ibm.com>