Move class chain length calculations into frontend
The length of the class chain associated to a class is relevant to the
JITServer AOT cache, as this quantity is used to construct RAMClass
chains that are sent from the client to the server. To support the
operation of the JITServer AOT cache without a local SCC, this length
must always be available.