Insert Null value check if array component type is unknown during compilation time
If the value being stored is NULL and the destination array
component is null restricted in runtime, a NPE is expected
to throw. Therefore, when the array component type is not
known to be identity type in compilation time, a NULLCHK
on store value is required.

Also add two test cases to test storing NULL into null restricted
array, and update the class name in the tests to primitive class
to reflect more accurately the type of the class.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>