Add numberOfElements parameter to getArrayletLayout()
0-length arrays have Discontiguous layout.
Currently detection of 0-length arrays based on dataSizeInBytes.
There is not good enough for 0-stride arrays obviously (have
dataSizeInBytes equal 0). All 0-stride arrays should be treated as
Contiguous with exception of 0-length 0-stride should be Discontiguous.
Passing of numberOfElements helps to solve this problem, dataSizeInBytes
is not required to be passed any more.

Signed-off-by: Dmitri Pivkine <Dmitri_Pivkine@ca.ibm.com>