Do not skip InjectedInvoker class in getCallerClass and getStackClass
This patch fixes both eclipse-openj9/openj9#14553 and
eclipse-openj9/openj9#18245. When OJDK MHs are enabled for JDK11, the
Hidden attribute for InjectedInvoker classes caused getStackClass and
getCallerClass to return the incorrect class. The solution is to
not iterate over InjectedInvoker classes despite them having the
J9AccMethodFrameIteratorSkip modifier set due to the Hidden attribute.

Closes: eclipse-openj9/openj9#14553 eclipse-openj9/openj9#18245
Signed-off-by: Nathan Henderson <nathan.henderson@ibm.com>