Jenkins: Always publish buildInfo to Artifactory
In the orignal change to add support for multiple
Artifactory servers, including servers behind a VPN
(eclipse-openj9#8817), I made a note that buildInfo cannot be published
to VPN'd servers because while the artifacts are pushed
from the node, the buildInfo is pushed from the Controller
and therefore would fail to push if the Controller could
not see the Artifactory server behind the VPN. Since that
change, we've move the UNB VPN connection directly onto our
Controller node, this means we can now publish buildInfo to
the UNB server. This essentially means that as of today, we
can push buildInfo to all servers.

Note we still need the Artifactory vpn config info for other
parts of the code.

Also note that the if condition  written prior to
this change was incorrect so we actually have never published
buildInfo since 8817 went in.

Supersedes eclipse-openj9#18203

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>