Use defining symrefs from defining map for <eaEscapeHelper> call
TR_EscapeAnalysisTools::processSymbolReferences uses a BitVector to
accumulate the symbol reference numbers for symbols that are to be
loaded on an <eaEscapeHelper> call.  In walking through the set of
defining symbols associated with a particular local symbol, the code was
adding the reference number for the original symbol to the BitVector of
symbols to be loaded for the <eaEscapeHelper> call rather than the
reference numbers for the associated defining symbols.  That can lead
Escape Analysis to miss potential escapes of references through induce
OSR blocks or peeked calls.