Allow loader data without local offsets
Class chain offsets do not necessarily need to be present when fetching
class loader data from the client in getClassRecord. We only really care
about the name of the loader at this point, since that is what is used
to create the server ClassLoader record. The server already tolerates
not having the local class chain offset cached - if it ever requires
that information, say during compilation, it will request the offset
from the client again.