Fix Tree Simplifier convertCurrentTimeMillis()
The convertCurrentTimeMillis() function was recreating a
currentTimeMillis() call into a new call target which would be replaced
in codegen with a Z specific instruction. To convert the results into
milliseconds the value has to be divided by a constant and the tree
was incorrectly converted into an ldiv with a call child that is not
anchored which allowed dead trees elimination to remove the treetop
resulting in the the ldiv to be moved in a way that would change the
semantic of the java method.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>