Use -fno-omit-frame-pointer on x86_64-mac JIT builds
A bug in libunwind on x86_64-macOS results in intermittent crashes from
libunwind if a C++ exception is thrown from
OMR::Compilation::failCompilation.  In order to work around the problem,
the JIT compiler needs to be compiled with -fno-omit-frame-pointer on
that combination of CPU and operating system.

That option will be used only on JDK 22 or later builds for now, while
the impact on compile-time and start-up performance is evaluated.