Apply idiomRecognition only to high frequency blocks
This commit attempts to reduce compilation overhead of
warm opt levels by not performing idiomRecognition on blocks
with a frequency lower than a specified threshold.
The threshold can be modified with
`-Xjit:idiomRecognitionFrequencyThresholdAtWarm=NNN`
A value of 0 means that idiomRecognition will be applied to
all blocks. A value of 10001 means the idiomRecognition will
not be applied at all, since 10000 is the largest possible block
frequency. Compilations at hot and above optimization levels
are not affected by this change.

Depends on eclipse/omr/#7162

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>