Fix SSL Test Failure due to time delays in SSL initialization
Due to time delays between SSL context initialization at the server and
and the client program exiting too soon, the SSL tests were failing.
The failure happens due to:
Required condition was not found: [Output match: Connected to a server]
As the client did not connect to the server as expected.

This fix reads the jitserver verbose log to confirm if the SSL context
is initialized before starting the client.

Closes: ##18308
Signed-off-by: SajinaKandy <sajina.kandy@ibm.com>