Update the fully mounted check in JVMTI Suspend/Resume Thread
During JVMTI suspend/resume thread operations, the current check to
determine if a thread is fully mounted is flaky. It depends on
J9VMThread->theadObject which might change after getVMThread due to
Java level changes, which are accepted as-is from the RI.

The check to determine if a thread is fully mounted is moved into
getVMThread. The check is updated to use the continuation state, which
is more stable since the Continuation implementation is maintained
within OpenJ9.

Related: #17865
Related: #17869
Related: #18370

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>