Prepare for value objects to be created with new bytecode
This change sets up changes to the JIT compiler that are needed to
permit value objects to be created with the new bytecode.  Previously
they could only be created with the aconst_init and withfield bytecode
instructions.  The new bytecode will be translated to the TR::New opcode
in IL, but existing optimizations that create value type instances will
continue to use the TR::newvalue opcode, so optimizations need to be
prepared for both possibilities for now.

Signed-off-by: Theresa Mammarella <Theresa.T.Mammarella@ibm.com>