Conform to spec for interface resolution for OJDK MHs
This patch fixes eclipse-openj9/openj9#14984.
For Java 10-, private interface lookups should fail with a
IncompatibleClassChangeError that is caught and wrapped in an
IllegalAccessException. For OJ9 MHs, the IllegalAccessException
is thrown in findInterface based on Java-code guards on method
modifiers. For OJDK MHs, the IncompatibleClassChangeError is
expected from a call to MethodHandleNatives.resolve and wrapped.

This patch ensures that the correct error is thrown for private
interfaces during method lookup for Java 11- with OJDK MHs.

Issues: eclipse-openj9/openj9#14984
Signed-off-by: Nathan Henderson <nathan.henderson@ibm.com>