Adjust heap size for CRIU restore
When we snapshot on a machine with higher physical memory and restore on one
with lower physical memory, the InstantOn JVM heap size will be larger than
a usual JVM run on the restore machine. And the momory usage will be abnormally
higher.

Update the extension with the correct physical memory limit and utilize the
existing softmax mechanism to solve the problem. Note we only update the softmax
if it is smaller than the Xmx size. If it is smaller than the
Xms size, then we upscale the softmx to match with the Xms size.

Related:
https://github.com/eclipse-openj9/openj9/issues/17596

Detailed behavior table:
https://github.com/eclipse-openj9/openj9/issues/18217#issuecomment-1799794593

Signed-off-by: Frank Kang frank.kang@ibm.com