Fix bug TR_LinkedListProfilerInfo
This commit fixes a bug in the code that adds a node to
TR_LinkedListProfilerInfo. Previously, if the linked list contained two
nodes, then adding a third one would result in a memory leak as the
third node was not added to the end of the list, but rather as the next
node of the first node. This meant that there was no longer any pointer
to the original second node in the list.

This commit ensures that a new node is always added to the end of the
list.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>