Add OPENJCEPLUS_SUPPORT flag in JPP tags
When building the OpenJCEPlus with Semeru OpenJDK, some java.base
packages need to be "exports" to the OpenJCEPlus module. But not
all the platforms are supported by OpenJCEPlus. For those not
supported platforms, export to OpenJCEPlus module will cause the
unknown module error. So, by adding the OPENJCEPLUS_SUPPORT flag
in JPP Tags, it will only allow the OpenJCEPlus required java.base
packages to be exported on the supported platforms, to avoid the
unknown module error.

Signed-off-by: Tao Liu <tao.liu@ibm.com>