Fix erroneously large section size in ROM walk
If a ROM class has no methods, the CP NAS section should have size zero.
Before, the section size would be calculated unconditionally based on
the location of the J9ROMCLASS_ROMMETHODS of the ROM class. When the
romMethods field is a null SRP, that would result in a negative pointer
difference that would then wrap to a very large U_32 value.

Signed-off-by: Christian Despres