Drop iprofiler buffers pre-checkpoint with -XX:+DebugOnRestore
Under -XX:+DebugOnRestore, in the pre-checkpoint run the IProfiler is
enabled; this is because the SCC cannot (yet) be used, so there is no
persistent profile info to draw from. However, the consequence of this
is higher RSS footprint.

As it turns out, the majority of the buffers processed by the IProfiler
is not needed to maintain throughput performance. This is because there
is a lot of duplicate data that comes in, as well as data that
represents highly execution paths.

Therefore, this commit drops 99.9% of the buffers. This ensures that
there the necessary data to ensure throughput performance is still
added to the IProfiler hash table.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>