Update JVM_VirtualThreadDisableSuspend
Enable JVM_VirtualThreadDisableSuspend in JDK22. A PR has been
opened to move it to JDK22 in the RI. Eventually, we will see this
method in JDK22.

The above method is invoked in the VirtualThread class. So, the
assertion in the stub will trigger and cause failures in the builds.

Leaving the stub empty should be fine. See #18671 for more details.

Related:
- #18671
- #18667
- #18663

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>