Fix wrong memoryPool names of memory usage in GCInfo
GCInfo contains a set of memory usages for all of memoryPools(both Heap
and NonHeap),java function buildGCInfo() retrieve a list of MemoryPools'
names and a list of Usages internally and generate the set of memory
usages, in some platforms(ex. MacOS) the order for the list of names and
for the list of usages might be different, the root cause is that we use
ManagementFactory.getMemoryPoolMXBeans() retrieve a copy of the list
MemoryPools and the copy list might be the different order than original
list for curtain platform, update to use the api
ExtendedMemoryMXBeanImpl.getInstance().getMemoryPoolMXBeans(false) to
retrieve original list to avoid misorder issue.

Signed-off-by: hulin <linhu@ca.ibm.com>