Support reimplemented MethodHandleProxies.asInterfaceInstance
In JDK22, MethodHandleProxies.asInterfaceInstance was reimplemented
using a more direct MethodHandle approach to improve performance.

This approach introduces hidden classes which have an interface for
its host class.

In this case, the hidden and host class don't need to exist in the
same package, and they can have different packages.

Two operations have been updated to accommodate the above changes:
[1] In ROMClassBuilder::handleAnonClassName, the host class's package
is not appended to the hidden class if the class name already has a
package.
[2] In internalDefineClass, hidden and host class are allowed to have
different package names if the host class is an interface.

[1] solves the NoClassDefFoundError seen in the below issues:
- #18694
- #18698
- #18699
- #18700
- #18701
- #18702

[2] solves the IllegalArgumentException, which is seen after the
NoClassDefFoundError is fixed.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>