Return true from Unsafe.shouldBeInitialized when called from <clinit>
When initialization of the class has started but not yet completed, and
shouldBeInitialized is called from the initializing thread, its result
should be true, but it has been returning false instead.

DirectMethodHandle.checkInitialized calls ensureClassInitialized and
then uses this method to determine whether initialization is incomplete
because it is still ongoing on the current thread. This fix should allow
that usage to work as intended.