Reduce counts only for implict SCC
PR #18356 introduced a change to reduce the counts of a method
whose class is not in the shared class cache (SCC), if SCC
is enabled and the load of that class happens during start-up.
This change improves performance of some short running benchmarks
from the DaCapo suite (when run with default options), but was
discovered to also cause a small (2-3%) throughput regression when
running the Daytrader benchmark on top of tWAS.
This commit restricts the change from PR #18356 to environments
that use an implicit SCC (by default, OpenJ9 creates an implict SCC
that is only allowed to to store bootstrap classes).

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>