Enable OpenJDK VarHandle call chain inlining
This commit contains the necessary changes to support
refining VM INL calls in VarHandle call chains, as well
as support for inlining the call targets down to the Unsafe
operations. This is necessary for improved OpenJDK VarHandle
performance.

VarHandles must be stored in static final fields for this to
work. VarHandle final fields will undergo folding during
ILGen by default in Java 17+ as we need this folding to
happen prior to inlining. It is not possible to modify
static final fields by conventional means in JDK17+.

In general, when VarHandle object is known, we can evaluate
the steps to obtain the target MethodHandle at compile time.
For that, we obtain the result of the following calls in
MethodHandleTransformer and InterpreterEmulator:
* Invokers.directVarHandleTarget
* VarHandle.asDirect
* Invokers.checkVarHandleGenericType

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>