Record/replay InterpreterEmulator constants
Inlining uses constant folding to help determine call targets, so it's
important for the corresponding IL (if any is generated) to be
consistent with what the inliner saw.

InterpreterEmulator now remembers constants for each TR_CallTarget that
must be folded consistently in the IL, and IL generation consults these
to repeat the same constant folding as needed.

This repeated folding is important whenever a value might be allowed to
be folded despite the possibility of a later change. It also allows
constants to be speculative by specifying the assumptions that are
necessary in order for the folding to be correct, and by informing the
IL generator of the locations where such assumptions have been made.
However, InterpreterEmulator does not yet do speculative folding.

This commit also removes a bunch of unused code, mostly related to class
lookahead, from TR_J9ByteCodeIlGenerator::loadStatic().