Accelerate String.indexOf(char) in JDK17+
Intrinsic methods for indexOf(char) in JITHelpers class is not used in JDK17+
because OpenJDK's java/lang/String class is used.
OpenJDK class library has intrinsic methods in java/lang/StringLatin1 and
java/lang/StringUTF16 classes. The logic of those intrinsic methods are
identical to those of JITHelpers class except that they are defined as
static methods. This commit adds intrinsic methods of StringLatin1 and
StringUTF16 to the recognized method table and update x/p/z/aarch64 codegen
to generate the specialized instruction sequence which already exists
for JITHelpers methods.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>