Do not add LambdaForm classes to SCC with OJDK MHs
This patch fixes eclipse-openj9/openj9#18757.
In JDK8 and JDK11 with OpenJDK MethodHandles, MH creation
uses defineAnonymousClass which passes a null class name
downstream for ROM class creation. The previous check and
flag-setting indicating that LambdaForms should not
be held in the SCC relies on a class name being non-null.
This patch pushes further downstream the flag-setting for
omitting LambdaForms from the SCC to ensure that the class
name is known (non-null) and available for comparison.

Issues: eclipse-openj9/openj9#18757
Signed-off-by: Nathan Henderson <nathan.henderson@ibm.com>